fix gc, simulation

+++ src/tools/common/nfs_client_impl.cpp
@@ -1,4 +1,4 @@
-# include "nfs_client_impl.h"
+﻿# include "nfs_client_impl.h"
 # include <dsn/internal/nfs.h>
 # include <queue>
 # include <boost/filesystem.hpp>
@@ -12,9 +12,23 @@ namespace dsn {
 			void* context)
 		{	
 			//dinfo("*** call RPC_NFS_COPY end, return (%d, %d) with %s", resp.offset, resp.size, err.to_string());
-			continue_copy(1); // consume left copy request
-
             copy_request_ex* reqc = (copy_request_ex*)context;
+
+			if (reqc->user_req->finished)
+				return;
+			task_ptr tsk = ::dsn::service::tasking::enqueue( // consume left copy request
+				CONTINUE_COPY,
+				this,
+				std::bind(
+				&nfs_client_impl::continue_copy,
+				this,
+				1
+				));
+			{
+				zauto_lock l(_lock);
+				_task_ptr_list.push_back(tsk);
+			}
+
 
 			if (err == ::dsn::ERR_SUCCESS)
 			{
@@ -28,22 +42,17 @@ namespace dsn {
 
 			if (err != ::dsn::ERR_SUCCESS)
 			{
-				{
-					zauto_lock l(_lock);
-					handle_fault(_file_path_map[reqc->copy_req.file_name], reqc->user_req, err);
-				}
+				handle_fault(_file_path_map[reqc->copy_req.file_name], reqc->user_req, err);
 				return;
 			}
 
 			if (resp.error != ::dsn::ERR_SUCCESS)
 			{
-				{
-					zauto_lock l(_lock);
-					handle_fault(_file_path_map[reqc->copy_req.file_name], reqc->user_req, err);
-				}
+				handle_fault(_file_path_map[reqc->copy_req.file_name], reqc->user_req, err);
 				return;
 			}
 
+			
             write_copy(reqc->user_req, resp);
 
             delete reqc;
@@ -108,45 +117,55 @@ namespace dsn {
 				{
 					it_resp->second.copy_response_map.insert(std::pair<uint64_t, copy_response>(resp.offset, resp));
 				}
+			}
 
-				::dsn::service::tasking::enqueue(
-					WRITE_FILE,
-					this,
-					std::bind(
-					&nfs_client_impl::write_file,
-					this,
-					req
-					));
+			if (req->finished)
+				return;
+
+			task_ptr tsk = ::dsn::service::tasking::enqueue(
+				WRITE_FILE,
+				this,
+				std::bind(
+				&nfs_client_impl::write_file,
+				this,
+				req
+				));
+			{
+				zauto_lock l(_lock);
+				_task_ptr_list.push_back(tsk);
 			}
         }
 
 		void nfs_client_impl::handle_fault(std::string file_path, user_request *req, error_code err)
 		{
-			derror("%s copy error", file_path.c_str());
-
-			auto it_failure = _file_failure_map.find(file_path); // add failure map
-			if (it_failure == _file_failure_map.end())
 			{
-				_file_failure_map.insert(std::pair<std::string, error_code>(file_path, err));
-			}
-			else
-				return;
+				zauto_lock l(_lock);
+				derror("%s copy error", file_path.c_str());
 
-			auto it_copy = _resp_copy_file_map.find(file_path); // remove copy map
+				auto it_failure = _file_failure_map.find(file_path); // add failure map
+				if (it_failure == _file_failure_map.end())
+				{
+					_file_failure_map.insert(std::pair<std::string, error_code>(file_path, err));
+				}
+				else
+					return;
 
-			if (it_copy != _resp_copy_file_map.end())
-			{
-				req->copy_request_count -= it_copy->second.copy_count;
-				req->copy_request_count += it_copy->second.finished_count;
-			}
-			else
-			{
-				req->copy_request_count -= _file_size_map[file_path] / _opts.max_buf_size + 1;
-			}
+				auto it_copy = _resp_copy_file_map.find(file_path); // remove copy map
 
-			if (it_copy != _resp_copy_file_map.end())
-			{
-				_resp_copy_file_map.erase(it_copy);
+				if (it_copy != _resp_copy_file_map.end())
+				{
+					req->copy_request_count -= it_copy->second.copy_count;
+					req->copy_request_count += it_copy->second.finished_count;
+				}
+				else
+				{
+					req->copy_request_count -= _file_size_map[file_path] / _opts.max_buf_size + 1;
+				}
+
+				if (it_copy != _resp_copy_file_map.end())
+				{
+					_resp_copy_file_map.erase(it_copy);
+				}
 			}
 
 			if (0 == req->copy_request_count)
@@ -157,25 +176,31 @@ namespace dsn {
 
 		void nfs_client_impl::handle_success(std::string file_path, user_request *req, error_code err)
 		{
-			auto it = _resp_copy_file_map.find(file_path);
-			it->second.finished_count++;
-
-			if (it->second.finished_count == it->second.copy_count)
+			int left_reqs;
 			{
-				if (boost::filesystem::exists(file_path))
-				{
-					file::close(_handles_map[file_path]);
-					_resp_copy_file_map.erase(it);
+				zauto_lock l(_lock);
+				auto it = _resp_copy_file_map.find(file_path);
+				if (it == _resp_copy_file_map.end())
+					return;
+				it->second.finished_count++;
 
-					auto it = _file_failure_map.find(file_path);
-					if (it != _file_failure_map.end())
+				if (it->second.finished_count == it->second.copy_count)
+				{
+					if (boost::filesystem::exists(file_path))
 					{
-						boost::filesystem::remove(file_path);
+						file::close(_handles_map[file_path]);
+						_resp_copy_file_map.erase(it);
+
+						auto it = _file_failure_map.find(file_path);
+						if (it != _file_failure_map.end())
+						{
+							boost::filesystem::remove(file_path);
+						}
 					}
 				}
-			}
 
-			auto left_reqs = --req->copy_request_count;
+				left_reqs = --req->copy_request_count;
+			}
 			if (0 == left_reqs)
 			{
 				handle_finish(file_path, req, err);
@@ -203,47 +228,70 @@ namespace dsn {
 					boost::filesystem::remove(it->first);
 				}
 				it++;
-				//_file_failure_map.erase(it++); // TODO: clean failure_map
 			}
+			
+			garbage_collect(req);
+		}
 
-			for (auto it = _handles_map.begin(); it != _handles_map.end();)
+		void nfs_client_impl::garbage_collect(user_request* req)
+		{
+			for (int i = 0; i < _task_ptr_list.size(); i++)
 			{
-				_handles_map.erase(it++);
+				if (_task_ptr_list[i] != nullptr)
+				{
+					_task_ptr_list[i]->cancel(true);
+				}
 			}
 
-			for (auto it = _file_size_map.begin(); it != _file_size_map.end();)
-			{
-				_file_size_map.erase(it++);
-			}
+			// wait a moment
+			std::this_thread::sleep_for(std::chrono::milliseconds(5000));
 
-			delete(req);
 			_file_size_map.clear();
-			//_file_failure_map.clear();
+			_file_failure_map.clear();
 			_resp_copy_file_map.clear();
+			_handles_map.clear();
+			_file_path_map.clear();
+			delete req;
 		}
 
 		void nfs_client_impl::internal_write_callback(error_code err, uint32_t sz, ::std::string file_path, user_request* req)
 		{
 			if (err != ERR_SUCCESS)
 			{
-				{
-					zauto_lock l(_lock);
-					handle_fault(file_path, req, err);
-				}
+				handle_fault(file_path, req, err);
 			}
 			if (err == ERR_SUCCESS)
 			{
+				bool success_flag = false;
 				{
 					zauto_lock l(_lock);
-
 					auto it = _file_failure_map.find(file_path);
 					if (it == _file_failure_map.end())
 					{
-						handle_success(file_path, req, err);
+						success_flag = true;
 					}
 				}
+				if (success_flag)
+					handle_success(file_path, req, err);
+			}
+
+			if (req->finished)
+				return;
+
+			task_ptr tsk = ::dsn::service::tasking::enqueue(
+				WRITE_FILE,
+				this,
+				std::bind(
+				&nfs_client_impl::write_file,
+				this,
+				req
+				));
+			{
+				zauto_lock l(_lock);
+				_task_ptr_list.push_back(tsk);
 			}
 		}
+
 		void nfs_client_impl::continue_copy(int done_count)
         {
             if (done_count > 0)
@@ -254,23 +302,42 @@ namespace dsn {
 				_concurrent_copy_request_count -= done_count;
             }
 
-            while (true)
+            copy_request_ex* req = nullptr;
             {
-                copy_request_ex* req = nullptr;
-                {
-                    zauto_lock l(_lock);
-                    if (_req_copy_file_queue.empty())
-                        break;
+                zauto_lock l(_lock);
 
-					if (_concurrent_copy_request_count >= _opts.max_concurrent_remote_copy_requests)
-						break;
-                    
-                    req = _req_copy_file_queue.front();
-                    _req_copy_file_queue.pop();
-					++_concurrent_copy_request_count;
-					
-					begin_copy(req->copy_req, req);
-                }
+				int empty_flag = 1;
+				if (!_req_copy_file_vector.empty() && _concurrent_copy_request_count < _opts.max_concurrent_remote_copy_requests)
+				{
+					for (int i = 0; i < _req_copy_file_vector.size(); i++)
+					{
+						if (_req_copy_file_vector[i].empty())
+							continue;
+						
+						req = _req_copy_file_vector[i].front();
+
+						auto it_failure = _file_failure_map.find(_file_path_map[req->copy_req.file_name]);
+						if (it_failure != _file_failure_map.end())
+						{
+							std::queue<copy_request_ex*> empty;
+							std::swap(_req_copy_file_vector[i], empty);
+							continue;
+						}
+
+						empty_flag = 0;
+						auto it_resp = _resp_copy_file_map.find(_file_path_map[req->copy_req.file_name]);
+						if (it_resp != _resp_copy_file_map.end())
+						{
+							if ((req->copy_req.offset - it_resp->second.current_offset) / _opts.max_buf_size > _opts.max_request_step) // exceed request bound for one file
+								continue;
+						}
+
+						_req_copy_file_vector[i].pop();
+						++_concurrent_copy_request_count;
+						begin_copy(req->copy_req, req);
+
+					}
+				}
             }
         }
 
@@ -315,18 +382,7 @@ namespace dsn {
 				}
 				if (_resp_copy_file_map.empty())
 					return;
-
-				::dsn::service::tasking::enqueue(
-					WRITE_FILE,
-					this,
-					std::bind(
-					&nfs_client_impl::write_file,
-					this,
-					req
-					));
 			}
-
-			
 		}
 
 		void nfs_client_impl::end_get_file_size(
@@ -382,6 +438,8 @@ namespace dsn {
 				else
 					req_size = size;
 
+				std::queue<copy_request_ex*> copy_request_ex_queue;
+				_req_copy_file_vector.push_back(copy_request_ex_queue);
 				for (;;) // send one file with multi-round rpc
 				{
                     copy_request_ex* req = new copy_request_ex;
@@ -399,7 +457,7 @@ namespace dsn {
 
 					{
 						zauto_lock l(_lock);
-						_req_copy_file_queue.push(req); // enqueue instead of sending it
+						_req_copy_file_vector[i].push(req); // enqueue instead of sending it
 					}
 
 					req_offset += req_size;
@@ -417,7 +475,19 @@ namespace dsn {
 				}
 			}
 
-			continue_copy();
+			task_ptr tsk = ::dsn::service::tasking::enqueue(
+				CONTINUE_COPY,
+				this,
+				std::bind(
+				&nfs_client_impl::continue_copy,
+				this,
+				0
+				));
+
+			{
+				zauto_lock l(_lock);
+				_task_ptr_list.push_back(tsk);
+			}
 		}
         		
 		void nfs_client_impl::begin_remote_copy(std::shared_ptr<remote_copy_request>& rci, aio_task_ptr nfs_task)
